pyproject.toml
[project]
name = "googleproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-genai>=1.14.0",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.0",
    "markdown>=3.8",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]
